+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Simple auto-scaling".
 - Tests for "Horse-racing hyperduals".
 - Tests for "Balanced ternary computer: encode".
+- Tests for "Extended hamming codes".
 
 ### Changed
 - Docker Node image from "14.16.1-alpine" to "node:16.14.2-alpine".